Completed parse_all_inputs method, updated readme.
parse_all_inputs now takes an input of the Words folder and a boolean
of whether to commit changes, and creates the dictionary database (as
previously advertised).

readme.md has been updated to show the current status and list the
existing challenges before an initial release can be completed.

Minor formatting changes to other code files.